Add support for paginating asset fetch
Under the hood this migrates our implementation from the Splunk Add-On
Builder REST "no-code" type to the Splunk Add-On Builder Python code
type.

To do this, I had to completely re-create the app, because the old code
was no longer fully importable into Splunk Add-On Builder.  The
pagination logic is in bin/input_module_assets.py.  The rest is mostly
add-on builder code.

🤞